// lib/tours/rendering/avatars/tour-avatar.types.ts
export type VideoCanvas = {
  width: number;
  height: number;
};

export type HeyGenAvatarAnchor = "bottom-right" | "bottom-left";

export type HeyGenAvatarSizePreset = {
  visibleWidthRatio: number;
  rightMargin: number;
  bottomMargin: number;
};

export const HEYGEN_AVATAR_SIZE_PRESETS = {
  small: {
    visibleWidthRatio: 0.33,
    rightMargin: 0,
    bottomMargin: 0,
  },
  medium: {
    visibleWidthRatio: 0.55,
    rightMargin: 0,
    bottomMargin: 0,
  },
  large: {
    visibleWidthRatio: 0.7,
    rightMargin: 0,
    bottomMargin: 0,
  },
} as const satisfies Record<string, HeyGenAvatarSizePreset>;

export type HeyGenAvatarSize = keyof typeof HEYGEN_AVATAR_SIZE_PRESETS;

export type HeyGenAvatarGenerationOptions = {
  aspectRatio: "16:9" | "9:16";
  fit: "contain";
  removeBackground: true;
  outputFormat: "webm";
  resolution: "720p" | "1080p";
  engineType?: "avatar_v" | "avatar_iv";
};

export type HeyGenAvatarPositioningInput = {
  anchor: HeyGenAvatarAnchor;
  rightMargin: number;
  bottomMargin: number;
  basis: "videoLayer" | "visibleBoundingBox";
  avatarWidth?: number | null;
  alphaThreshold?: number;
};

export type HeyGenAvatarResolvedPositioning = Omit<
  HeyGenAvatarPositioningInput,
  "alphaThreshold" | "avatarWidth"
> & {
  avatarWidth: number | null;
  alphaThreshold: number;
};
// ...
export type HeyGenAvatarAlphaAnalysis = {
  sourceWidth: number;
  sourceHeight: number;
  sampledFrameCount: number;
  alphaThreshold: number;
  medianBox: VisiblePixelBox;
  maxBox: VisiblePixelBox;
  transparentPadding: AvatarTransparentPadding;
  edgeTouchRate: AvatarEdgeTouchRate;
  cropRisk: AvatarCropRisk;
};

export type VisiblePixelBox = {
  x: number;
  y: number;
  width: number;
  height: number;
  right: number;
  bottom: number;
};

export type AvatarTransparentPadding = {
  left: number;
  right: number;
  top: number;
  bottom: number;
};

export type AvatarEdgeTouchRate = {
  left: number;
  right: number;
  top: number;
  bottom: number;
};

export type AvatarCropRisk = {
  level: "none" | "low" | "medium" | "high";
  reasons: string[];
};

export type HeyGenAvatarResolvedPlacement = {
  avatarWidth: number;
  anchor: HeyGenAvatarAnchor;
  rightMargin: number;
  bottomMargin: number;
  basis: "videoLayer" | "visibleBoundingBox";
  overlayX: string;
  overlayY: string;
};

export type HeyGenAvatarFfmpegOverlayOptions = {
  avatarInputCodec: "libvpx-vp9";
  backgroundFilter: string;
  avatarScaleFilter: string;
  overlayFilter: string;
  filterComplex: string;
  outputVideoCodec: "libx264";
  outputAudioCodec: "aac";
  preserveAlpha: true;
};

export type HeyGenAvatarOverlayPlan = {
  canvas: VideoCanvas;
  size: HeyGenAvatarSize;
  placement: HeyGenAvatarResolvedPlacement;
  ffmpeg: HeyGenAvatarFfmpegOverlayOptions;
};
// ...
export class TourAvatarError extends Error {
  constructor(
    message: string,
    readonly code:
      | "MISSING_HEYGEN_API_KEY"
      | "MISSING_HEYGEN_AVATAR_ID"
      | "MISSING_AVATAR_SOURCE"
      | "HEYGEN_CREATE_FAILED"
      | "HEYGEN_POLL_FAILED"
      | "HEYGEN_DOWNLOAD_FAILED"
      | "AVATAR_ALPHA_ANALYSIS_FAILED"
      | "AVATAR_LAYOUT_FAILED"
      | "AVATAR_VIDEO_UPLOAD_FAILED"
      | "AVATAR_METADATA_UPLOAD_FAILED"
      | "AVATAR_VIDEO_ASSET_CREATE_FAILED"
      | "AVATAR_METADATA_ASSET_CREATE_FAILED"
      | "FRAME_CHECK_FAILED"
  ) {
    super(message);
    this.name = "TourAvatarError";
  }
}